Refactor the `getModuleSemester` function that is unrelated to `lessonIndices` into `utils/index`
